Move utils to module level up
