+++ rail_bot/bot/subscription/unsubscribe_handler.py
@@ -15,7 +15,7 @@ from telegram.ext import CallbackContext, CallbackQueryHandler, CommandHandler
 from rail_bot.bot.job_manager import JobManager
 from rail_bot.bot.service.subscription_service import Travel
 from rail_bot.bot.subscription.common import UNSUBSCRIBE
-from rail_bot.bot.utils import parse_time
+from rail_bot.utils import parse_time
 
 logger = logging.getLogger(__name__)
